Successfully implemented Repository level mapping Unit Tests

+++ AdventureWorks.Domain/HumanResources/EmployeeDepartmentHistory.cs
@@ -6,7 +6,7 @@ namespace AdventureWorks.Domain.HumanResources
 	public class EmployeeDepartmentHistory : EntityBase
 	{
 		public int BusinessEntityId { get; set; }
-		public short DepartmentId { get; set; }
+		public int DepartmentId { get; set; }
 		public byte ShiftId { get; set; }
 		public DateTime StartDate { get; set; }
 		public DateTime? EndDate { get; set; }
